requirements.txt
jax
jaxlib
platypus-opt
pandas
seaborn
hydroeval
shapely
pyproj
rtree
fiona
geopandas
requests
tqdm
netCDF4
xarray